+++ packages/system-functions/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stoker-platform/system-functions
 
+## 0.5.96
+
+### Patch Changes
+
+- feat: update Algolia search filters to match web app filters
+
 ## 0.5.95
 
 ### Patch Changes